feat: add optional support for `log` facade
This commit adds optional support for the `log`
facade by implementing a concrete adapter that
wraps an inner `Logger`.

This approach introduces the least amount of
changes, respects the maintainers wishes to
introduce additive changes that do not modify the
internal Logger trait, and the accompanying macros,
and delegates logging to the internal `Logger`.

The adapter, initialization function, and `log`
macros are made available via the cargo feature
"facade".

We enable the "kv" feature on "log" to enable users/
callers enrich the log facade `Record` with key-value
pairs, and then transmit them to the internal `Record`